تغیر long به int برای قیمت 1400/12/08

+++ CR.Core/DTOs/Payment/RedirectToPaymentForWalletChargeDto.cs
@@ -3,6 +3,6 @@
     public class RedirectToPaymentForWalletChargeDto
     {
         public string transactionNumber { get; set; }
-        public long price { get; set; }
+        public int price { get; set; }
     }
 }
